remove special tests, update for PHP 8
