--- phpunit-system.xml.dist
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<phpunit backupGlobals="false"
-         backupStaticAttributes="false"
-         colors="true"
-         convertErrorsToExceptions="true"
-         convertNoticesToExceptions="true"
-         convertWarningsToExceptions="true"
-         processIsolation="false"
-         stopOnFailure="false"
-         bootstrap="tests/bootstrap-system.php"
->
-    <testsuites>
-        <testsuite name="PHP JSON Web Token Test Suite">
-            <file>./tests/CachedKeySetTest.php</file>
-        </testsuite>
-    </testsuites>
-</phpunit>
